[refactor]PMA-65: move createTask method from columnsService to tasksService
