+++ src/app/boards/components/column/column.component.html
@@ -27,7 +27,7 @@
       [task]="task"
     ></app-task>
   </mat-card-content>
-  <button mat-icon-button>
+  <button mat-icon-button (click)="onClickCreateTask()">
     <mat-icon color="primary">note_add</mat-icon>
     Add a Task
   </button>
